ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi>=0.135.1",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.78.0",
    "pytest>=9.0.2",
]
